+++ internal/core/services/scheduler_manager/scheduler_manager.go
@@ -281,5 +281,5 @@ func (s *SchedulerManager) newSchedulerInfo(ctx context.Context, scheduler *enti
 	if err != nil {
 		return nil, fmt.Errorf("failing in couting game rooms in %s state: %s", game_room.GameStatusTerminating, err)
 	}
-	return entities.NewSchedulerInfo(scheduler.Name, scheduler.Game, scheduler.State, int64(ready), int64(occupied), int64(creating), int64(terminating)), nil
+	return entities.NewSchedulerInfo(scheduler.Name, scheduler.Game, scheduler.State, ready, occupied, creating, terminating), nil
 }
